♻️ refactor(client): move AskSageClient to separate module
- extract AskSageClient from models.py to client.py for modularity
- improve code organization and separation of concerns

♻️ refactor(models): import AskSageClient from client module

- remove duplicated client logic from models.py
- update import to use new asksage_proxy.client location
